Added method setAddressTextAndQuery
This new method has one address parameter. It changes the address text to the provided address and then queries the input. Using _handleChangeText(address) and setListViewDisplayed(true), it also shows the list of results automatically. This gives more flexibility to developers who would like to autofill and query input programmatically.